Rename cpus.ts -> cpu.ts
