+++ ui/src/components/cpu.ts
